feat: Make creation of Kubernetes resources togglable (#8)
The storage module currently requires a Kubernetes cluster to provision a bucket, which shouldn't be a requirement.

* This PR adds the ability to opt out of Kubernetes resource creation.
* Still defaults to provisioning a Kubernetes config map.

Note:
This will trigger the Kubernetes config map to be replaced after upgrade (which is fine)
This is because the Terraform resource address will change (adding an index). The use of a moved-block is not implemented, as we still maintain backwards compatibility with >=0.13.2.

What to expect:
* The resource module.x.kubernetes_config_map.main will be deleted.
* The resource module.x.kubernetes_config_map[0].main will be created.

The new resource is identical. The replacement should be brief, and will not impact running pods.

Author:    Sindre Lindstad <35308882+sindrel@users.noreply.github.com>